Separate AST definition in Types.hs into AST.hs
