+++ hs-src/Language/Egison/Desugar.hs
@@ -23,6 +23,7 @@ import           Data.Set              (Set)
 import qualified Data.Set              as S
 
 import           Language.Egison.Types
+import           Language.Egison.AST
 
 desugarTopExpr :: EgisonTopExpr -> EgisonM EgisonTopExpr
 desugarTopExpr (Define name expr)   = Define name <$> desugar expr
